change typography exports
