--- src/components/Typography/Typography1.css
@@ -1,3 +1,0 @@
-.semibold-title{
-    font-size: 32px;
-}
