--- src/components/Typography/Typography1.jsx
@@ -1,15 +1,0 @@
-// Monserat semi-bold
-
-import React from 'react'
-
-import './Typography1.css'
-
-const Typography1 = ({text}) => {
-    return (
-    <div >
-        <h1 className = "semibold-title"> {text} </h1>
-    </div>
-    )
-}
-
-export default Typography1
