--- src/components/Typography/Typography2.css
@@ -1,4 +1,0 @@
-.medium-title{
-    font-weight: 500;
-    font-size: 24px;
-}
