--- src/components/Typography/Typography2.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import './Typography2.css'
-
-const Typography2 = ({text}) => {
-    return (
-    <div >
-        <h1 className = "medium-title"> {text} </h1>
-    </div>
-    )
-}
-
-export default Typography2
