--- src/components/Typography/Typography3.css
@@ -1,5 +1,0 @@
-.regularItalic-title {
-    font-style: italic;
-    font-size: 18px;
-    color: #aaaaaa;
-}
