--- src/components/Typography/Typography3.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import './Typography3.css'
-
-const Typography3 = ({text}) => {
-    return (
-    <div >
-        <h1 className = "regularItalic-title"> {text} </h1>
-    </div>
-    )
-}
-
-export default Typography3
