--- src/components/Typography/Typography4.css
@@ -1,5 +1,0 @@
-.bold-titles > h1{
-    text-transform: uppercase;
-    font-size: 22px;
-    font-weight: bold;
-}
