--- src/components/Typography/Typography4.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import './Typography4.css'
-
-const Typography4 = ({text}) => {
-    return (
-    <div  className = "bold-titles">
-        <h1> {text} </h1>
-    </div>
-    )
-}
-
-export default Typography4
