--- src/components/Typography/Typography5.css
@@ -1,5 +1,0 @@
-.semiBold-titles > h1{
-    text-transform: uppercase;
-    font-size: 22px;
-    color: #aaaaaa;
-}
