--- src/components/Typography/Typography5.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import './Typography5.css'
-
-const Typography5 = ({text}) => {
-    return (
-    <div  className = "semiBold-titles">
-        <h1> {text} </h1>
-    </div>
-    )
-}
-
-export default Typography5
