+++ src/components/Typography/index.js
@@ -1,0 +1,7 @@
+export {
+    Typography1,
+    Typography2,
+    Typography3,
+    Typography4,
+    Typography5,
+} from './Typography';
